SmartQ&A/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
langchain-anthropic==0.1.0
langchain-text-splitters==0.0.1
chromadb==0.4.18
pypdf==3.17.4
docx2txt==0.8
pydantic==2.5.0
